requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
billiard==4.2.1
boto3==1.40.8
botocore==1.40.8
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
colorlover==0.3.0
comm==0.2.3
cufflinks==0.17.3
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
ipython==8.37.0
ipywidgets==8.1.7
jedi==0.19.2
jmespath==1.0.1
jupyterlab_widgets==3.0.15
kombu==5.5.4
matplotlib-inline==0.1.7
narwhals==2.1.1
numpy==2.2.6
opencv-contrib-python==4.7.0.72
packaging==25.0
pandas==2.3.1
parso==0.8.4
pillow==11.3.0
plotly==6.3.0
prompt_toolkit==3.0.51
pure_eval==0.2.3
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.47.2
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
wcwidth==0.2.13
widgetsnbextension==4.0.14
xlrd==2.0.1
pymysql
torch